requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
anyio==3.6.1; python_full_version >= '3.6.2'
certifi==2022.6.15.1; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==2.1.1; python_version >= '3.6'
click==8.1.3; python_version >= '3.7'
cryptography==38.0.1; python_version >= '3.6'
deprecated==1.2.13; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
dlib==19.24.0
face-recognition-models==0.3.0
face-recognition==1.3.0
fastapi==0.83.0
gcloud==0.18.3
googleapis-common-protos==1.56.4; python_version >= '3.7'
httplib2==0.20.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
idna==3.3; python_version >= '3.5'
jwcrypto==1.3.1; python_version >= '3.0'
numpy==1.23.3; python_version >= '3.8'
oauth2client==4.1.3
opencv-python==4.6.0.66
pandas==1.4.4
pillow==9.2.0
protobuf==4.21.5; python_version >= '3.7'
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pydantic==1.10.2; python_version >= '3.7'
pyparsing==3.0.9; python_version >= '3.1'
pyrebase4==4.5.0
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-jwt==3.3.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2022.2.1
requests-toolbelt==0.9.1
requests==2.28.1; python_version >= '3.7' and python_version < '4'
rsa==4.9; python_version >= '3.6' and python_version < '4'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0; python_version >= '3.7'
starlette==0.19.1; python_version >= '3.6'
typing-extensions==4.3.0; python_version < '3.10'
urllib3==1.26.12; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4'
wrapt==1.14.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
